consolidate hooks into one and simplify the components
